\name{unit.hypersphere.volume}
\alias{unit.hypersphere.volume}

\title{
  Helps When Calculating Densities
}
\description{
  Finds the volume of a hypersphere in R^D with radius one.
}
\usage{
unit.hypersphere.volume(D)
}

\arguments{
  \item{D}{
    the number of dimensions that the hypersphere extends into
  }
}
\value{
  the volume of the unit hypersphere
}
\references{
  http://en.wikipedia.org/wiki/N-sphere
}

\keyword{ multivariate }
